--- src/main/java/org/folio/innreach/domain/service/AgencyService.java
@@ -1,9 +1,0 @@
-package org.folio.innreach.domain.service;
-
-import org.folio.innreach.dto.CentralServerAgenciesDTO;
-
-public interface AgencyService {
-
-  CentralServerAgenciesDTO getAllAgencies();
-
-}
