+++ src/main/java/org/folio/innreach/domain/service/CentralServerConfigurationService.java
@@ -1,0 +1,12 @@
+package org.folio.innreach.domain.service;
+
+import org.folio.innreach.dto.CentralServerAgenciesDTO;
+import org.folio.innreach.dto.CentralServerItemTypesDTO;
+
+public interface CentralServerConfigurationService {
+
+  CentralServerAgenciesDTO getAllAgencies();
+
+  CentralServerItemTypesDTO getAllItemTypes();
+
+}
